Catagorize each endpoint to by resource type
