+++ README.md
@@ -24,23 +24,31 @@ robinhood.getQuoteBySymbol('BAC')
 *NOTE: `id` means a single instrument ID (e.g. `ebab2398-028d-4939-9f1d-13bf38f81c50`), 
 `symbol` means a single instrument symbol (e.g. `FB`)*
 
-- `get100MostPopular()`      
-    *return*: A list of instrument ids under the 100 most popular tag      
+- `instrument`
+  - `getInstrument(id)`
+      &nbsp;&nbsp;*return*: An instrument object
     
-- `getInstrument(id)`      
-    *return*: An instrument object      
-    
-- `getInstruments(id_list)`      
-    *`id_list`* List of instrument ids      
-    *return*: List of instrument objects      
-    
-- `getRatings(id)`    
-    *return*: Ratings object about instrument with id      
-    
-- `getQuoteBySymbol(symbol)`
-- `getQuote(id)`
-- `getPopularity(id)`
-- `getPopularities(id_list)`
-- `getInstrumentBySymbol(symbol)`
-- `getHistoricalBySymbol(symbol_list, interval)`      
-  *`interval` time interval, could be `week`, `day`...
+  - `getInstruments(id_list)`
+      &nbsp;&nbsp;*`id_list`* List of instrument ids
+      &nbsp;&nbsp;*return*: List of instrument objects
+
+  - `getInstrumentBySymbol(symbol)`
+
+- `midlands`
+  - `getRatings(id)`
+      &nbsp;&nbsp;*return*: Ratings object about instrument with id
+
+  - `get100MostPopular()`
+      &nbsp;&nbsp;*return*: A list of instrument ids under the 100 most popular tag
+
+- `quote`
+  - `getQuoteBySymbol(symbol)`
+  - `getQuote(id)`
+
+- `popularity`
+  - `getPopularity(id)`
+  - `getPopularities(id_list)`
+
+- `historical`
+  - `getHistoricalBySymbol(symbol_list, interval)`
+    &nbsp;&nbsp;*`interval` time interval, could be `week`, `day`...
