--- lib/crawler/robinhood/index.js
@@ -1,2 +1,0 @@
-'use strict';
-
