--- lib/rating/rating_base.js
@@ -1,13 +1,0 @@
-'use strict';
-
-class Rating {
-  /**
-   * Get ratings interface
-   * @param {Array<string>} [symbols] A list of symbol to get ratings of
-   */
-  getRatings(symbols) {
-    throw new Error('Not implemented');
-  }
-}
-
-module.exports = Rating;
